move more legacy sidebar code
